requirements.txt
appdirs==1.4.4
arrow==0.17.0
backcall==0.2.0
binaryornot==0.4.4
boto3==1.16.13
botocore==1.19.13
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
idna==2.10
importlib-metadata==1.7.0
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.10.0
MarkupSafe==1.1.1
parso==0.7.1
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
poyo==0.5.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
pynamodb==4.3.3
python-dateutil==2.8.1
python-slugify==4.0.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
stevedore==3.2.0
text-unidecode==1.3
traitlets==5.0.5
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
